+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next
-- Adds a possibility to set Options > Application Language and Application Region for a `TestAction` on a scheme by @paciej00
+
+### Added
+
+- Adds a possibility to set Options > Application Language and Application Region for a `TestAction` on a scheme [#1055](https://github.com/tuist/tuist/pull/1055) by [@paciej00](https://github.com/paciej00)
 
 ### Changed